Fix memory leak in test helper
