+++ c_code/zopfli.c
@@ -3465,99 +3465,3 @@ void ZopfliDeflate(const ZopfliOptions *options, int btype, int final,
     }
 }
 
-void single_test(const unsigned char *in, int btype, int blocksplitting, int blocksplittingmax)
-{
-    ZopfliOptions options;
-
-    // Configure options
-    options.verbose = 1;
-    options.verbose_more = 0; // Reduce internal verbose output
-    options.numiterations = 15;
-    options.blocksplitting = blocksplitting;
-    options.blocksplittinglast = 0;
-    options.blocksplittingmax = blocksplittingmax;
-
-    unsigned char *out = 0;
-    size_t outsize = 0;
-    unsigned char bp = 0;
-    size_t insize = strlen(in);
-
-    // Perform compression
-    ZopfliDeflate(&options, btype, 1, (const unsigned char *)in, insize, &bp, &out, &outsize);
-}
-
-void run_all_tests(const unsigned char *in)
-{
-    single_test(in, 2, 1, 15); // Dynamic Huffman
-    single_test(in, 1, 1, 15); // Fixed Huffman
-    single_test(in, 0, 1, 15); // Uncompressed
-
-    // Test with different block splitting settings
-    single_test(in, 2, 0, 15); // No block splitting
-    single_test(in, 2, 1, 5);  // Limited splits
-    single_test(in, 2, 1, 0);  // No splits
-    single_test(in, 2, 1, 1);  // No splits
-    single_test(in, 2, 1, 50); // No splits
-    single_test(in, 2, 1, 30); // More splits allowed
-}
-
-unsigned char *read_stdin_to_bytes(size_t *out_size)
-{
-    // Initial buffer size
-    size_t buffer_size = 1024;
-    size_t total_read = 0;
-
-    // Allocate initial buffer
-    unsigned char *buffer = malloc(buffer_size);
-    if (buffer == NULL)
-    {
-        fprintf(stderr, "Memory allocation failed\n");
-        return NULL;
-    }
-
-    // Read stdin in chunks
-    int ch;
-    while ((ch = getchar()) != EOF)
-    {
-        // Resize buffer if needed
-        if (total_read >= buffer_size)
-        {
-            buffer_size *= 2;
-            unsigned char *new_buffer = realloc(buffer, buffer_size);
-            if (new_buffer == NULL)
-            {
-                fprintf(stderr, "Memory reallocation failed\n");
-                free(buffer);
-                return NULL;
-            }
-            buffer = new_buffer;
-        }
-
-        // Store the character
-        buffer[total_read++] = (unsigned char)ch;
-    }
-
-    // Set the output size
-    *out_size = total_read;
-    return buffer;
-}
-
-int main()
-{
-    size_t bytes_size;
-    unsigned char *bytes = read_stdin_to_bytes(&bytes_size);
-
-    if (bytes == NULL)
-    {
-        fprintf(stderr, "Failed to read stdin\n");
-        return 1;
-    }
-
-    // Run tests
-    run_all_tests(bytes);
-
-    // Free allocated memory
-    free(bytes);
-
-    return 0;
-}
